requirements.txt
fastapi
uvicorn[standard]
pydicom
pillow
numpy
pytest
